+++ app/common/containers/pages/SignInPage/redux.js
@@ -5,6 +5,7 @@ import { createSessionToken } from 'redux/auth';
 import { fetchUserData } from 'redux/user';
 import { login } from 'redux/session';
 import { CLIENT_ID } from 'config';
+import error_messages from 'helpers/errors';
 
 export const onSubmit = ({ email, password }) => (dispatch, getState) =>
 dispatch(createSessionToken({
@@ -17,14 +18,6 @@ dispatch(createSessionToken({
 .then((action) => {
   if (action.error) {
     const { message = 'SMS not send. Try later' } = action.payload.response.error;
-
-    const error_messages = {
-      'User blocked': 'user_blocked',
-      'You reached login attempts limit. Try again later': 'reached_max_attemps',
-      'Identity, password combination is wrong.': 'emailOrPasswordMismatch',
-      'SMS not send. Try later': 'resentOtp',
-      'Sending OTP timeout. Try later.': 'otp_timeout',
-    };
 
     if (message) {
       throw new SubmissionError({
