+++ app/common/helpers/errors.js
@@ -1,0 +1,9 @@
+const error_messages = {
+  'User blocked': 'user_blocked',
+  'You reached login attempts limit. Try again later': 'reached_max_attemps',
+  'SMS not send. Try later': 'resentOtp',
+  'Sending OTP timeout. Try later.': 'otp_timeout',
+  'Identity, password combination is wrong.': 'emailOrPasswordMismatch',
+};
+
+export default error_messages;
